Fixing PR comments

+++ server/controllers/invites.js
@@ -9,6 +9,8 @@ const mongoose = require('mongoose');
 const config = require('../config');
 const tokenEncrypt = require('../utils/password');
 const email = require('../utils/email');
+const inviteTemplate = require('../assets/templates/email/invite.template.html');
+const slug = require('slug')
 
 const Invite = require('../models/invites');
 const Company = require('../models/companies');
@@ -28,24 +30,15 @@ exports.create = (req, res, next) => {
 
             const url = `https://${req.get('host')}/invite/${invite['token']}`;
             const message = invite['message'] ? `<p style="text-align: left;width: 100%; color: #000; margin-bottom: 25px;">${invite['message']}</p>` : '<!-- Message -->';
-            invite['html'] =
-              `<div style="font-family: sans-serif;padding: 75px 25px;background: #f1f5f4;width: 100%;display:-webkit-flex;display:-ms-flexbox;display:flex;-webkit-box-orient:horizontal;-webkit-box-direction:normal;-webkit-flex-direction:row;-ms-flex-direction:row;flex-direction:row;-webkit-box-pack:center;-webkit-justify-content:center;-ms-flex-pack:center;justify-content:center;">
-                <div style="padding: 25px 25px;background: #fff;width: 100%;max-width: 500px; text-align: center;padding-bottom: 15px; margin: 0 auto;">
-                  <h3 style="text-align: left;width: 100%; color: #000;">Hi,</h3>
-                  ${message}
-                  <p style="text-align: left;width: 100%; color: #000;" margin-bottom: 35px;>Please click on the button below to create an account:</p>
-                  <a href="${url}" style="display:inline-block;padding: 12px 20px;background: #5078f1;color: #fff;font-size:12px;font-weight: 500;border-radius: 5px; text-decoration:none;" target="_blank">Create an account</a>
-                  <p style="text-align: center;margin-top: 50px; color: #000; opacity: .4; width: 100%;">Ambrosus - all rights reserved, 2018.</p>
-                </div>
-              </div>`;
+            invite['html'] = inviteTemplate.replace(/@url/g, url).replace(/@message/g, message);
             invite['company'] = company;
           });
 
           Invite.insertMany(invites).then(invites => {
             invites.forEach(invite => {
               const invitation = JSON.parse(JSON.stringify(invite));
-              invitation.subject = 'Create an account, Ambrosus Dashboard';
-              invitation.from = 'no-reply@ambrosus.com';
+              invitation.subject = `${user.full_name} invited you to join ${user.company.title} Dasbhoard`;
+              invitation.from = `no-reply@${slug(user.company.title)}.com`;
               email.send(invitation)
                 .then(sent => console.log('Email sent'))
                 .catch(error => console.log('Email send error: ', error));
@@ -68,8 +61,9 @@ exports.create = (req, res, next) => {
 
 exports.delete = (req, res, next) => {
   const ids = req.body.ids || [];
+  const userId = req.session.user._id;
 
-  Invite.deleteMany({ _id: { $in: ids } })
+  Invite.deleteMany({ _id: { $in: ids }, from: userId })
     .then(deleted => {
       req.status = 200;
       req.json = { message: 'Successfuly deleted', data: deleted };
@@ -81,6 +75,7 @@ exports.getAll = (req, res, next) => {
   const company = req.params.company;
 
   Invite.find({ company })
+    .populate('from')
     .select('-__v')
     .then(invites => {
       req.status = 200;
